created utils folder and extracted input validation functions
