+++ src/app/chatbot/page.module.css
@@ -1,7 +1,7 @@
 .chatContainer {
   align-items: flex-start;
   background-color: lightgray;
-  border-radius: 3%;
+
   display: flex;
   flex-direction: column;
   gap: 15px;
